now when the coin is picked it moves to its empty socket

+++ Motor2D/j1Coin.cpp
@@ -51,7 +51,23 @@ bool Coin::Awake(pugi::xml_node&)
 bool Coin::Update(float dt)
 {
 	BROFILER_CATEGORY("Coin Update", Profiler::Color::Cyan);
+	if (moving)
+	{
+		fPoint dPos;
+		dPos.x = distanceTo.x/(0.3 / dt);
+		dPos.y = distanceTo.y/(0.3 / dt);
 
+		virtualPosition.x -= dPos.x;
+		virtualPosition.y -= dPos.y;
+		if (pos_relCam <= goingTo.x && virtualPosition.y <= goingTo.y)
+		{
+			j1Player* player = (j1Player*)App->entityManager->getPlayer();
+			player->score += 500;
+			player->coins[id - 1] = true;
+			moving = false;
+			App->entityManager->DeleteEntity(this);
+		}
+	}
 
 	return true;
 }
@@ -71,18 +87,28 @@ bool Coin::CleanUp()
 }
 void Coin::OnCollision(Collider* c1, Collider* c2)
 {
-	if (c2->type == COLLIDER_PLAYER && !App->entityManager->getPlayer()->dead)
+	if (c2->type == COLLIDER_PLAYER && !App->entityManager->getPlayer()->dead && !moving)
 	{
 		App->audio->PlayFx(earn_coin_fx, 0);
-		App->entityManager->DeleteEntity(this);
 
 		j1Player* player = (j1Player*)App->entityManager->getPlayer();
 		if (!player->coins[id - 1])
 		{
-			player->score += 500;
-			player->coins[id - 1] = true;
+			moveTo(player->coins_pos[id - 1].x, player->coins_pos[id - 1].y);
+		}
+		else
+		{
+			App->entityManager->DeleteEntity(this);
 		}
 	}
+}
+
+void Coin::moveTo(int x, int y)
+{
+	goingTo = { x,y };
+	distanceTo.x = pos_relCam - x;
+	distanceTo.y = position.y - y;
+	moving = true;
 }
 
 bool Coin::Load(pugi::xml_node&)
